Begin implementing Markdown support
