--- demo/posts/Post-20200825-The Spanish Inquisition.html
@@ -1,41 +1,0 @@
-
-<p>
-Pero las llamas son peligrosas. Si usted ve una llama donde hay gente﻿ nadando, usted gritar: ¡Cuidado! ¡Llamas! Why is it the world never remembered the name of Johann Gambolputty de von Ausfern-schplenden-schlitter-crasscrenbon-fried-digger-dangle-dongle-dungle-burstein-Von-knacker-thrasher-apple-banger-horowitz-ticolensic-grander-knotty-spelltinkle-grandlich-grumblemeyer-spelterwasser-kurstlich-himbleeisen-bahnwagen-gutenabend-bitte-ein-nürnburger-bratwurstle-gerspurten-mit-zwei-macheluber-hundsfut-gumberaber-shoenendanker-kalbsfleisch-mittler-aucher von Hautkopft of Ulm?
-</p>
-
-<img align="right" src="assets/images/cleese.jpg">
-<p>
-
-Get on with it! The purpose of foreplay is to cause the vagina to lubricate so that the penis can penetrate more easily. I cut down trees, I skip and jump, I like to press wildflowers. I put own womens' clothing, and hang around in bars. I object to all this sex on the television. I mean, I keep falling off!
-
-</p>
-<p>
-We interrupt this program to annoy you and make things generally irritating. Look at them, bloody Catholics, filling the bloody world up with bloody people they can't afford to bloody feed. Oh and Jenkins, apparently your mother died this morning. I'm afraid I have no choice but to sell you all for scientific experiments. That's what being a Protestant's all about. That's why it's the church for me. That's why it's the church for anyone who respects the individual and the individual's right to decide for him or herself. When Martin Luther nailed his protest up to the church door in fifteen-seventeen, he may not have realised the full significance of what he was doing, but four hundred years later, thanks to him, my dear, I can wear whatever I want on my John Thomas...
-</p>
-<p>
-Shut your festering gob, you tit! Your type really makes me puke you vacuous, toffy-nosed, malodorous pervert! This morning, shortly after 11:00, comedy struck this little house on Dibley Road. Sudden, violent comedy. Hegel is arguing that the reality is merely an a priori adjunct of non-naturalistic ethics, Kant via the categorical imperative is holding that ontologically it exists only in the imagination, and Marx claims it was offside. Nobody expects the Spanish Inquisition!
-
-</p>
-<p>
-It's a Mr. Death, dear. He's here about the reaping. If a sperm is wasted God gets quite irate. All right, settle down. Settle down... Now, before I begin the lesson, will those of you who are playing in the match this afternoon move your clothes down onto the lower peg immediately after lunch, before you write your letter home, if you're not getting your hair cut, unless you've got a younger brother who is going out this weekend as the guest of another boy, in which case, collect his note before lunch, put it in your letter after you've had your hair cut, and make sure he moves your clothes down onto the lower peg for you.
-
-</p>
-<p>
-In 1945, peace broke out. It was the end of the Joke. Joke warfare was banned at a special session of the Geneva Convention, and in 1950 the last remaining copy of the joke was laid to rest here in the Berkshire countryside, never to be told again. Well, er, yes Mr. Anchovy, but you see your report here says that you are an extremely dull person. You see, our experts describe you as an appallingly dull fellow, unimaginative, timid, lacking in initiative, spineless, easily dominated, no sense of humour, tedious company and irrepressibly drab and awful. And whereas in most professions these would be considerable drawbacks, in chartered accountancy, they're a positive boon.
-
-</p>
-<p>
-Not necessarily. I could be arguing in my spare time. Oh, waiter! This conversation isn't very good. Is your wife a goer, eh? Know what I mean? Know what I mean? Nudge, nudge! Know what I mean? Say no more! It's not pining, it's passed on! This parrot is no more! It has ceased to be! It's expired and gone to meet its maker! This is a late parrot! It's a stiff! Bereft of life, it rests in peace! If you hadn't nailed it to the perch, it would be pushing up the daisies! It's metabolic processes are now history! He's off the twig! He's kicked the bucket, he's shuffled off the mortal coil, rung down the curtain and joined the choir invisible. This is an ex-parrot!
-
-</p>
-<p>
-If we took the bones out, it wouldn't be crunchy, would it? I told you to lay off the beans, you whore! And finally, a wafer thin mint. Well, had I got as far as the penis entering the vagina? And Dinsdale says 'I hear you've been a naughty boy, Clement', and he splits me nostrils open, saws me leg off and pulls me liver out. And I tell him 'My name's not Clement', and then he loses his temper and nails my head to the floor.
-
-</p>
-<p>
-I mean, because we are members of the Protestant Reformed Church, which successfully challenged the autocratic power of the Papacy in the mid-sixteenth century, we can wear little rubber devices to prevent issue. What's wrong with a kiss, boy? Hmm? Why not start her off with a nice kiss? You don't have to go leaping straight for the clitoris like a bull at a gate. Give her a kiss, boy.
-
-</p>
-<p>
-Ah, I see you have the machine that goes ping. This is my favorite. You see we lease it back from the company we sold it to and that way it comes under the monthly current budget and not the capital account. We use only the finest baby frogs, dew picked and flown from Iraq, cleansed in finest quality spring water, lightly killed, and then sealed in a succulent Swiss quintuple smooth treble cream milk chocolate envelope and lovingly frosted with glucose. We have to go. Uhm... I'm having rather heavy period. In 1945, peace broke out. It was the end of the Joke. Joke warfare was banned at a special session of the Geneva Convention, and in 1950 the last remaining copy of the joke was laid to rest here in the Berkshire countryside, never to be told again.
-</p>
