--- demo/posts/Post-20220319-A dead parrot.html
@@ -1,279 +1,0 @@
-<p>
-The cast:
-MR. PRALINE
-John Cleese
-SHOP OWNER
-Michael Palin
-</p>
-
-</p>
-The sketch:
-A customer enters a pet shop.
-</p>
-
-<p>
-Mr. Praline: 'Ello, I wish to register a complaint.
-(The owner does not respond.)
-<p></p>
-Mr. Praline: 'Ello, Miss?
-
-<p></p>
-Owner: What do you mean "miss"?
-<p></p>
-
-Mr. Praline: (pause)I'm sorry, I have a cold. I wish to make a complaint!
-<p></p>
-
-Owner: We're closin' for lunch.
-<p></p>
-
-Mr. Praline: Never mind that, my lad. I wish to complain about this parrot what I purchased not half an hour ago from this very boutique.
-<p></p>
-
-Owner: Oh yes, the, uh, the Norwegian Blue...What's,uh...What's wrong with it?
-<p></p>
-
-Mr. Praline: I'll tell you what's wrong with it, my lad. 'E's dead, that's what's wrong with it!
-<p></p>
-
-Owner: No, no, 'e's uh,...he's resting.
-<p></p>
-
-Mr. Praline: Look, matey, I know a dead parrot when I see one, and I'm looking at one right now.
-<p></p>
-
-Owner: No no he's not dead, he's, he's restin'! Remarkable bird, the Norwegian Blue, idn'it, ay? Beautiful plumage!
-<p></p>
-
-Mr. Praline: The plumage don't enter into it. It's stone dead.
-<p></p>
-
-Owner: Nononono, no, no! 'E's resting!
-<p></p>
-
-Mr. Praline: All right then, if he's restin', I'll wake him up! (shouting at the cage) 'Ello, Mister Polly Parrot! I've got a lovely fresh cuttle fish for you if you show...
-<p></p>
-
-(owner hits the cage)
-<p></p>
-
-Owner: There, he moved!
-<p></p>
-
-Mr. Praline: No, he didn't, that was you hitting the cage!
-<p></p>
-
-Owner: I never!!
-<p></p>
-
-Mr. Praline: Yes, you did!
-<p></p>
-
-Owner: I never, never did anything...
-<p></p>
-
-Mr. Praline: (yelling and hitting the cage repeatedly) 'ELLO POLLY!!!!! Testing! Testing! Testing! Testing! This is your nine o'clock alarm call!
-<p></p>
-
-(Takes parrot out of the cage and thumps its head on the counter. Throws it up in the air and watches it plummet to the floor.)
-<p></p>
-
-Mr. Praline: Now that's what I call a dead parrot.
-<p></p>
-
-Owner: No, no.....No, 'e's stunned!
-<p></p>
-
-Mr. Praline: STUNNED?!?
-<p></p>
-
-Owner: Yeah! You stunned him, just as he was wakin' up! Norwegian Blues stun easily, major.
-<p></p>
-
-Mr. Praline: Um...now look...now look, mate, I've definitely 'ad enough of this. That parrot is definitely deceased, and when I purchased it not 'alf an hour ago, you assured me that its total lack of movement was due to it bein' tired and shagged out following a prolonged squawk.
-<p></p>
-
-Owner: Well, he's...he's, ah...probably pining for the fjords.
-<p></p>
-
-Mr. Praline: PININ' for the FJORDS?!?!?!? What kind of talk is that?, look, why did he fall flat on his back the moment I got 'im home?
-<p></p>
-
-Owner: The Norwegian Blue prefers keepin' on it's back! Remarkable bird, id'nit, squire? Lovely plumage!
-<p></p>
-
-Mr. Praline: Look, I took the liberty of examining that parrot when I got it home, and I discovered the only reason that it had been sitting on its perch in the first place was that it had been NAILED there.
-<p></p>
-
-(pause)
-<p></p>
-
-Owner: Well, o'course it was nailed there! If I hadn't nailed that bird down, it would have nuzzled up to those bars, bent 'em apart with its beak, and VOOM! Feeweeweewee!
-<p></p>
-
-Mr. Praline: "VOOM"?!? Mate, this bird wouldn't "voom" if you put four million volts through it! 'E's bleedin' demised!
-<p></p>
-
-Owner: No no! 'E's pining!
-<p></p>
-
-Mr. Praline: 'E's not pinin'! 'E's passed on! This parrot is no more! He has ceased to be! 'E's expired and gone to meet 'is maker! 'E's a stiff! Bereft of life, 'e rests in peace! If you hadn't nailed 'im to the perch 'e'd be pushing up the daisies! 'Is metabolic processes are now 'istory! 'E's off the twig! 'E's kicked the bucket, 'e's shuffled off 'is mortal coil, run down the curtain and joined the bleedin' choir invisible!! THIS IS AN EX-PARROT!!
-<p></p>
-
-(pause)
-<p></p>
-
-Owner: Well, I'd better replace it, then. (he takes a quick peek behind the counter) Sorry squire, I've had a look 'round the back of the shop, and uh, we're right out of parrots.
-<p></p>
-
-Mr. Praline: I see. I see, I get the picture.
-<p></p>
-
-Owner: (pause) I got a slug.
-<p></p>
-
-(pause)
-<p></p>
-
-Mr. Praline: Pray, does it talk?
-<p></p>
-
-Owner: Nnnnot really.
-<p></p>
-
-Mr. Praline: WELL IT'S HARDLY A BLOODY REPLACEMENT, IS IT?!!???!!?
-<p></p>
-
-Owner: N-no, I guess not. (gets ashamed, looks at his feet)
-<p></p>
-
-Mr. Praline: Well.
-<p></p>
-
-(pause)
-<p></p>
-
-Owner: (quietly) D'you.... d'you want to come back to my place?
-<p></p>
-
-Mr. Praline: (looks around) Yeah, all right, sure.
-<p></p>
-
-Alternate ending:
-<p></p>
-Mr. Praline: (sweet as sugar) Pray, does it talk?
-<p></p>
-
-Owner: Nnnnot really.
-<p></p>
-
-Mr. Praline: WELL IT'S HARDLY A BLOODY REPLACEMENT, IS IT?!!???!!?
-<p></p>
-
-Owner: Look, if you go to my brother's pet shop in Bolton, he'll replace the parrot for you.
-<p></p>
-
-Mr. Praline: Bolton, eh? Very well.
-<p></p>
-
-(The customer leaves.)
-<p></p>
-
-(The customer enters the same pet shop. The owner is putting on a false moustache.)
-<p></p>
-
-Mr. Praline: This is Bolton, is it?
-<p></p>
-
-Owner: (with a fake mustache) No, it's Ipswitch.
-<p></p>
-
-Mr. Praline: (looking at the camera) That's inter-city rail for you.
-<p></p>
-
-(Mr. Praine goes to the train station. He addresses a man standing behind a desk marked "Complaints".)
-<p></p>
-
-Mr. Praline: I wish to complain, British-Railways Person.
-<p></p>
-
-Attendant: I DON'T HAVE TO DO THIS JOB, YOU KNOW!!!
-<p></p>
-
-Mr. Praline: I beg your pardon...?
-<p></p>
-
-Attendant: I'm a qualified brain surgeon! I only do this job because I like being my own boss!
-<p></p>
-
-Mr. Praline: Excuse me, this is irrelevant, isn't it?
-<p></p>
-
-Attendant: Yeah, well it's not easy to pad these python files out to 150 lines, you know.
-<p></p>
-
-Mr. Praline: Well, I wish to complain. I got on the Bolton train and found myself deposited here in Ipswitch.
-<p></p>
-
-Attendant: No, this is Bolton.
-<p></p>
-
-Mr. Praline: (to the camera) The pet shop man's brother was lying!!
-<p></p>
-
-Attendant: Can't blame British Rail for that.
-<p></p>
-
-Mr. Praline: In that case, I shall return to the pet shop!
-<p></p>
-
-He does.
-<p></p>
-
-Mr. Praline: I understand this IS Bolton.
-<p></p>
-
-Owner: (still with the fake mustache) Yes?
-<p></p>
-
-Mr. Praline: You told me it was Ipswitch!
-<p></p>
-
-Owner: ...It was a pun.
-<p></p>
-
-Mr. Praline: (pause) A PUN?!?
-<p></p>
-
-Owner: No, no...not a pun...What's that thing that spells the same backwards as forwards?
-<p></p>
-
-Mr. Praline: (Long pause) A palindrome...?
-<p></p>
-
-Owner: Yeah, that's it!
-<p></p>
-
-Mr. Praline: It's not a palindrome! The palindrome of "Bolton" would be "Notlob"!! It don't work!!
-<p></p>
-
-Owner: Well, what do you want?
-<p></p>
-
-Mr. Praline: I'm not prepared to pursue my line of inquiry any longer as I think this is getting too silly!
-<p></p>
-
-Sergeant-Major: Quite agree, quite agree, too silly, far too silly... (takes customer by the arm) Come on, you, you've got to go do another sketch now! Come on... (he walks off stage left, followed by the director and cameramen, leaving the owner alone on the set)
-<p></p>
-
-Owner: (to the audience) Well! I never wanted to do this in the first place. I wanted to be... a lumberjack!
-<p></p>
-
-(he takes off his white lab coat to reveal a checkered shirt and suspenders under it)
-<p></p>
-
-Floating down the mighty rivers of British Columbia! With my best girl by my side!...
-<p></p>
-<p>
-Sketch continues into the Lumberjack Song.
-</p>
