--- demo/posts/Post-20220510-How to add new post.html
@@ -1,21 +1,0 @@
-<p>
-To add a new post, just add a file in the <b>posts</b> folder with the correct
-naming convention.
-</p>
-
-<h3>Post file format</h3>
-<p>
-Post files should follow the rule <b>&lt;LAYOUT&gt;&lt;DATE&gt;&lt;TITLE&gt;.html</b>.
-</p>
-<p> <b>&lt;LAYOUT&gt;</b> is the name of a layout file under the current theme/layout folder. </p>
-<p> <b>&lt;DATE&gt;</b> is the post creation date in the format YYYYMMDD. </p>
-<p> <b>&lt;TITLE&gt;</b> is the title of the post. </p>
-
-<h3>Limitations</h3>
-<p>
-WHile the naming convention is simple and practical, it imposes a technical
-limitation because some characters are not allowed as file names. I'll probably
-come up with a syntax expressing such characters so they are substituted when
-generating the pages.
-</p>
-
