--- demo/posts/Post-20220811-Hot to create a Layout file.html
@@ -1,32 +1,0 @@
-
-<p>
-Layout files are templates on how to generate posts. For this simple demo site,
-this is the source code for the template used for the posts.
-</p>
-
-<pre>
-&lt;html&gt;
-  {{include &quot;include/header.html&quot;}}
-  &lt;body&gt;
-    &lt;div class=&quot;content&quot;&gt;
-      {{include &quot;include/top.html&quot;}}
-      &lt;h2&gt;{{post.title}}&lt;/h2&gt;
-      &lt;time datetime=&quot;{{post.year}}-{{post.month}}-{{post.day}}:00:00+02:00&quot;&gt;
-        {{post.year}}-{{post.month_name}}-{{post.day}}
-      &lt;/time&gt;
-      &lt;article&gt;{{post.body}}&lt;/article&gt;
-      &lt;a href=&quot;index.html&quot;/&gt;Back&lt;/a&gt;
-      {{include &quot;include/footer.html&quot;}}
-    &lt;/div&gt;
-  &lt;/body&gt;
-&lt;/html&gt;
-</pre>
-
-<p>
-It {{include}} the top.html with renders the top banner, and the navigation links.
-Then it displays the title of the post with the {{post.title}} tag. Then the post
-creation date with {{post.year}}, {{post.month_name}} and {{post.day}} tags.
-Then, finally it renders the content of the post with the {{post.body}} tag,
-right before {{include}}'ing the footer
-</p>
-
