--- demo/posts/Post-20220820-No Markup.html
@@ -1,19 +1,0 @@
-<p>
-If you took some time to read the source of the files inder the posts folder
-you'll notice they are plain html files.
-</p>
-
-<h2>Markdown support</h2>
-
-<p>
- I don't like the idea of only having mardown as an input for posts. Yes, it's
- probably the thing you'll use the most but sometimes you need to add extra
- stuff to a post and then you are stuck to the markdown.
-</p>
-
-<p>
-My plan is to support markdown expansion between {{markdown}} {{endmarkdown}}
-tags. So the generator knows that it must expand that specific block while
-ignoring anything else. This makes possible for users to actually write custom
-HTML code when they need to.
-</p>
