+++ src/markdown.h
@@ -1,0 +1,12 @@
+#ifndef MARKDOWN
+#define MARKDOWN
+
+#include <string>
+struct ParseContext;
+
+namespace markdown
+{
+  std::string markdownToHtml(char* sourceStart, char* sourceEnd);
+}
+
+#endif  //MARKDOWN
